+++ mjx/mujoco/mjx/_src/solver.py
@@ -22,6 +22,7 @@ from mujoco.mjx._src import smooth
 from mujoco.mjx._src import support
 # pylint: disable=g-importing-member
 from mujoco.mjx._src.dataclasses import PyTreeNode
+from mujoco.mjx._src.types import ConeType
 from mujoco.mjx._src.types import Data
 from mujoco.mjx._src.types import DisableBit
 from mujoco.mjx._src.types import Model
@@ -45,8 +46,12 @@ class _Context(PyTreeNode):
     cost: constraint + Gauss cost
     prev_cost: cost from previous iter
     solver_niter: number of solver iterations
+    active: active (quadratic) constraints            (nefc,)
+    fri: friction of regularized cone                 (num(con.dim > 1), 6)
+    dm: regularized constraint mass                   (num(con.dim > 1))
+    u: friction cone (normal and tangents)            (num(con.dim > 1), 6)
+    h: cone hessian                                   (num(con.dim > 1), 6, 6)
   """
-
   qacc: jax.Array
   qfrc_constraint: jax.Array
   Jaref: jax.Array  # pylint: disable=invalid-name
@@ -59,6 +64,11 @@ class _Context(PyTreeNode):
   cost: jax.Array
   prev_cost: jax.Array
   solver_niter: jax.Array
+  active: jax.Array
+  fri: jax.Array
+  dm: jax.Array
+  u: jax.Array
+  h: jax.Array
 
   @classmethod
   def create(cls, m: Model, d: Data, grad: bool = True) -> '_Context':
@@ -66,6 +76,15 @@ class _Context(PyTreeNode):
     # TODO(robotics-team): determine nv at which sparse mul is faster
     ma = support.mul_m(m, d, d.qacc)
     nv_0 = jp.zeros(m.nv)
+    fri = 0.0
+    if m.opt.cone == ConeType.ELLIPTIC:
+      friction = d.contact.friction[d.contact.dim > 1]
+      dim = d.contact.dim[d.contact.dim > 1]
+      mu = friction[:, 0] / jp.sqrt(m.opt.impratio)
+      fri = jp.concatenate((mu[:, None], friction), axis=1)
+      for condim in (3, 4, 6):
+        fri = fri.at[dim == condim, condim:].set(0)
+
     ctx = _Context(
         qacc=d.qacc,
         qfrc_constraint=d.qfrc_constraint,
@@ -79,8 +98,13 @@ class _Context(PyTreeNode):
         cost=jp.inf,
         prev_cost=0.0,
         solver_niter=0,
+        active=0.0,
+        fri=fri,
+        dm=0.0,
+        u=0.0,
+        h=0.0,
     )
-    ctx = _update_constraint(d, ctx)
+    ctx = _update_constraint(m, d, ctx)
     if grad:
       ctx = _update_gradient(m, d, ctx)
       ctx = ctx.replace(search=-ctx.Mgrad)  # start with preconditioned gradient
@@ -106,24 +130,68 @@ class _LSPoint(PyTreeNode):
   @classmethod
   def create(
       cls,
+      m: Model,
       d: Data,
       ctx: _Context,
       alpha: jax.Array,
       jv: jax.Array,
       quad: jax.Array,
       quad_gauss: jax.Array,
+      uu: jax.Array,
+      v0: jax.Array,
+      uv: jax.Array,
+      vv: jax.Array,
   ) -> '_LSPoint':
     """Creates a linesearch point with first and second derivatives."""
     # roughly corresponds to CGEval in mujoco/src/engine/engine_solver.c
 
     # TODO(robotics-team): change this to support friction constraints
-    active = ((ctx.Jaref + alpha * jv) < 0).at[:d.ne + d.nf].set(True)
-    quad = jax.vmap(jp.multiply)(quad, active)  # only active
-    quad_total = quad_gauss + jp.sum(quad, axis=0)
+    cost, deriv_0, deriv_1 = 0.0, 0.0, 0.0
+    quad_total = quad_gauss
 
-    cost = alpha * alpha * quad_total[2] + alpha * quad_total[1] + quad_total[0]
-    deriv_0 = 2 * alpha * quad_total[2] + quad_total[1]
-    deriv_1 = 2 * quad_total[2] + (quad_total[2] == 0) * mujoco.mjMINVAL
+    if m.opt.cone == ConeType.ELLIPTIC:
+      mu, u0 = ctx.fri[:, 0], ctx.u[:, 0]
+      n = u0 + alpha * v0
+      tsqr = uu + alpha * (2 * uv + alpha * vv)
+      t = jp.sqrt(tsqr)  # tangential force
+
+      bottom_zone = ((tsqr <= 0) & (n < 0)) | ((tsqr > 0) & ((mu * n + t) <= 0))
+      middle_zone = (tsqr > 0) & (n < (mu * t)) & ((mu * n + t) > 0)
+
+      # quadratic cost for equality, friction, limits, frictionless contacts
+      dim1 = d.contact.efc_address[d.contact.dim == 1]
+      nefl = d.ne + d.nf + d.nl
+      active = ((ctx.Jaref + alpha * jv) < 0).at[:d.ne + d.nf].set(True)
+      active = active.at[nefl:].set(False).at[dim1].set(active[dim1])
+      quad_efld = jax.vmap(jp.multiply)(quad, active)
+      quad_total += jp.sum(quad_efld, axis=0)
+      # elliptic bottom zone: quadratic cost
+      efc_elliptic = d.contact.efc_address[d.contact.dim > 1]
+      quad_c = jax.vmap(jp.multiply)(quad[efc_elliptic], bottom_zone)
+      quad_total += jp.sum(quad_c, axis=0)
+      # elliptic middle zone
+      t += (t == 0) * mujoco.mjMINVAL
+      tsqr += (tsqr == 0) * mujoco.mjMINVAL
+      n1 = v0
+      t1 = (uv + alpha * vv) / t
+      t2 = vv / t - (uv + alpha * vv) * t1 / tsqr
+      dm = ctx.dm * middle_zone
+      nmt = n - mu * t
+      cost = 0.5 * jp.sum(dm * jp.square(nmt))
+      deriv_0 = jp.sum(dm * nmt * (n1 - mu * t1))
+      deriv_1 = jp.sum(dm * (jp.square(n1 - mu * t1) - nmt * mu * t2))
+    elif m.opt.cone == ConeType.PYRAMIDAL:
+      active = ((ctx.Jaref + alpha * jv) < 0).at[:d.ne + d.nf].set(True)
+      quad = jax.vmap(jp.multiply)(quad, active)  # only active
+      quad_total += jp.sum(quad, axis=0)
+    else:
+      raise NotImplementedError(f'unsupported cone type: {m.opt.cone}')
+
+    alpha_sq = alpha * alpha
+    cost += alpha_sq * quad_total[2] + alpha * quad_total[1] + quad_total[0]
+    deriv_0 += 2 * alpha * quad_total[2] + quad_total[1]
+    deriv_1 += 2 * quad_total[2] + (quad_total[2] == 0) * mujoco.mjMINVAL
+
     return _LSPoint(alpha=alpha, cost=cost, deriv_0=deriv_0, deriv_1=deriv_1)
 
 
@@ -159,34 +227,95 @@ def _while_loop_scan(cond_fun, body_fun, init_val, max_iter):
   return jax.lax.scan(_fun, init, None, length=max_iter)[0][0]
 
 
-def _update_constraint(d: Data, ctx: _Context) -> _Context:
+def _update_constraint(m: Model, d: Data, ctx: _Context) -> _Context:
   """Updates constraint force and resulting cost given latst solver iteration.
 
   Corresponds to CGupdateConstraint in mujoco/src/engine/engine_solver.c
 
   Args:
+    m: model defining constraints
     d: data which contains latest qacc and smooth terms
     ctx: current solver context
 
   Returns:
     context with new constraint force and costs
   """
-  # TODO(robotics-team): add friction constraints
+  if m.opt.cone == ConeType.PYRAMIDAL:
+    # ne/nf constraints are always active, rest are non-negative constraints
+    active = (ctx.Jaref < 0).at[:d.ne + d.nf].set(True)
+    efc_force = d.efc_D * -ctx.Jaref * active
+    cost = 0.5 * jp.sum(d.efc_D * ctx.Jaref * ctx.Jaref * active)
+    dm, u, h = 0.0, 0.0, 0.0
+  elif m.opt.cone == ConeType.ELLIPTIC:
+    friction = d.contact.friction[d.contact.dim > 1]
+    efc_address = d.contact.efc_address[d.contact.dim > 1]
+    dim = d.contact.dim[d.contact.dim > 1]
+    slice_fn = jax.vmap(lambda x: jax.lax.dynamic_slice(ctx.Jaref, (x,), (6,)))
+    u = slice_fn(efc_address) * ctx.fri
+    mu, n, t = ctx.fri[:, 0], u[:, 0], jax.vmap(math.norm)(u[:, 1:])
 
-  # only count active constraints
-  active = (ctx.Jaref < 0).at[:d.ne + d.nf].set(True)
+    # bottom zone: quadratic
+    bottom_zone = ((t <= 0) & (n < 0)) | ((t > 0) & ((mu * n + t) <= 0))
+    active = (ctx.Jaref < 0).at[:d.ne + d.nf].set(True)
+    adr_i, adr_j = [], []
+    for i, (condim, addr) in enumerate(zip(dim, efc_address)):
+      adr_i.extend(range(addr, addr + condim))
+      adr_j.extend([i] * condim)
+    active = active.at[jp.array(adr_i)].set(bottom_zone[jp.array(adr_j)])
+    efc_force = d.efc_D * -ctx.Jaref * active
+    cost = 0.5 * jp.sum(d.efc_D * ctx.Jaref * ctx.Jaref * active)
 
-  efc_force = d.efc_D * -ctx.Jaref * active
+    # middle zone: cone
+    middle_zone = (t > 0) & (n < (mu * t)) & ((mu * n + t) > 0)
+    dm = d.efc_D[efc_address] / jp.maximum(
+        mu * mu * (1 + mu * mu), mujoco.mjMINVAL
+    )
+    nmt = n - mu * t
+    cost += 0.5 * jp.sum(dm * nmt * nmt * middle_zone)
+    # tangent and friction for middle zone:
+    force = -dm * nmt * mu * middle_zone
+    force_fri = -force / (t + ~middle_zone * mujoco.mjMINVAL)
+    force_fri = force_fri[:, None] * u[:, 1:] * friction
+    efc_force = efc_force.at[efc_address].add(force)
+    efc_adr, adr_i, adr_j = [], [], []
+    for i, (condim, addr) in enumerate(zip(dim, efc_address)):
+      efc_adr.extend(range(addr + 1, addr + condim))
+      adr_i.extend([i] * (condim - 1))
+      adr_j.extend(range(condim - 1))
+    efc_adr, adr_i, adr_j = jp.array(efc_adr), jp.array(adr_i), jp.array(adr_j)
+    efc_force = efc_force.at[efc_adr].add(force_fri[(adr_i, adr_j)])
+
+    # cone hessian
+    h = 0.0
+    if m.opt.solver == SolverType.NEWTON:
+      t = jp.maximum(t, mujoco.mjMINVAL)
+      # h = mu*N/T^3 * U*U'
+      ttt = jp.maximum(t * t * t, mujoco.mjMINVAL)
+      h = jax.vmap(lambda x, y: x * jp.outer(y, y.T))(mu * n / ttt, u)
+      # add to diagonal: (mu^2 - mu*N/T) * I
+      h += jax.vmap(lambda x: x * jp.eye(6, 6))(mu * mu - mu * n / t)
+      # set first row: (1, -mu/T * U)
+      h_0 = jax.vmap(lambda mu, t, u: jp.append(1, -mu / t * u[1:]))(mu, t, u)
+      h = h.at[:, 0].set(h_0).at[:, :, 0].set(h_0)
+      # pre and post multiply by diag(mu, friction), scale by Dm
+      h *= jax.vmap(lambda d, f: d * jp.outer(f, f.T))(dm, ctx.fri)
+      # only cone constraints
+      h = jax.vmap(jp.multiply)(h, middle_zone)
+  else:
+    raise NotImplementedError(f'unsupported cone type: {m.opt.cone}')
+
   qfrc_constraint = d.efc_J.T @ efc_force
   gauss = 0.5 * jp.dot(ctx.Ma - d.qfrc_smooth, ctx.qacc - d.qacc_smooth)
-  cost = 0.5 * jp.sum(d.efc_D * ctx.Jaref * ctx.Jaref * active) + gauss
-
   ctx = ctx.replace(
       qfrc_constraint=qfrc_constraint,
       gauss=gauss,
-      cost=cost,
+      cost=cost + gauss,
       prev_cost=ctx.cost,
       efc_force=efc_force,
+      active=active,
+      dm=dm,
+      u=u,
+      h=h,
   )
 
   return ctx
@@ -213,8 +342,17 @@ def _update_gradient(m: Model, d: Data, ctx: _Context) -> _Context:
   if m.opt.solver == SolverType.CG:
     mgrad = smooth.solve_m(m, d, grad)
   elif m.opt.solver == SolverType.NEWTON:
-    active = (ctx.Jaref < 0).at[: d.ne + d.nf].set(True)
-    h = (d.efc_J.T * d.efc_D * active) @ d.efc_J
+    if m.opt.cone == ConeType.ELLIPTIC:
+      cm = jp.diag(d.efc_D * ctx.active)
+      efc_address = d.contact.efc_address[d.contact.dim > 1]
+      dim = d.contact.dim[d.contact.dim > 1]
+      # set efc of cone H along diagonal
+      for i, (condim, addr) in enumerate(zip(dim, efc_address)):
+        h_cone = ctx.h[i, :condim, :condim]
+        cm = cm.at[addr:addr+condim, addr:addr+condim].add(h_cone)
+      h = d.efc_J.T @ cm @ d.efc_J
+    else:
+      h = (d.efc_J.T * d.efc_D * ctx.active) @ d.efc_J
     h = support.full_m(m, d) + h
     h_ = jax.scipy.linalg.cho_factor(h)
     mgrad = jax.scipy.linalg.cho_solve(h_, grad)
@@ -256,8 +394,28 @@ def _linesearch(m: Model, d: Data, ctx: _Context) -> _Context:
   ))
   quad = jp.stack((0.5 * ctx.Jaref * ctx.Jaref, jv * ctx.Jaref, 0.5 * jv * jv))
   quad = (quad * d.efc_D).T
+  uu, v0, uv, vv = 0.0, 0.0, 0.0, 0.0
+  if m.opt.cone == ConeType.ELLIPTIC:
+    mask = d.contact.dim > 1
+    # complete vector quadratic (for bottom zone)
+    efc_con, efc_fri = [], []
+    for condim, addr in zip(d.contact.dim[mask], d.contact.efc_address[mask]):
+      efc_con.extend([addr] * (condim - 1))
+      efc_fri.extend(range(addr + 1, addr + condim))
+    quad = quad.at[jp.array(efc_con)].add(quad[jp.array(efc_fri)])
 
-  point_fn = lambda a: _LSPoint.create(d, ctx, a, jv, quad, quad_gauss)
+    # rescale to make primal cone circular
+    jv_fn = jax.vmap(lambda x: jax.lax.dynamic_slice(jv, (x,), (6,)))
+    efc_elliptic = d.contact.efc_address[mask]
+    v = jv_fn(efc_elliptic) * ctx.fri
+    uu = jp.sum(ctx.u[:, 1:] * ctx.u[:, 1:], axis=1)
+    v0 = v[:, 0]
+    uv = jp.sum(ctx.u[:, 1:] * v[:, 1:], axis=1)
+    vv = jp.sum(v[:, 1:] * v[:, 1:], axis=1)
+
+  point_fn = lambda a: _LSPoint.create(
+      m, d, ctx, a, jv, quad, quad_gauss, uu, v0, uv, vv
+  )
 
   def cond(ctx: _LSContext) -> jax.Array:
     done = ctx.ls_iter >= m.opt.ls_iterations
@@ -274,21 +432,34 @@ def _linesearch(m: Model, d: Data, ctx: _Context) -> _Context:
     hi_next = point_fn(hi.alpha - hi.deriv_0 / hi.deriv_1)
     mid = point_fn(0.5 * (lo.alpha + hi.alpha))
 
-    # we swap lo/hi if:
-    # 1) they are not correctly at a bracket boundary (e.g. lo.deriv_0 > 0), OR
-    # 2) if moving to next or mid narrows the bracket
-    swap_lo_next = (lo.deriv_0 > 0) | (lo.deriv_0 < lo_next.deriv_0)
-    lo = jax.tree_util.tree_map(lambda x, y: jp.where(swap_lo_next, y, x), lo, lo_next)
-    swap_lo_mid = (mid.deriv_0 < 0) & (lo.deriv_0 < mid.deriv_0)
-    lo = jax.tree_util.tree_map(lambda x, y: jp.where(swap_lo_mid, y, x), lo, mid)
-
-    swap_hi_next = (hi.deriv_0 < 0) | (hi.deriv_0 > hi_next.deriv_0)
-    hi = jax.tree_util.tree_map(lambda x, y: jp.where(swap_hi_next, y, x), hi, hi_next)
-    swap_hi_mid = (mid.deriv_0 > 0) & (hi.deriv_0 > mid.deriv_0)
-    hi = jax.tree_util.tree_map(lambda x, y: jp.where(swap_hi_mid, y, x), hi, mid)
-
-    swap = swap_lo_next | swap_lo_mid | swap_hi_next | swap_hi_mid
-
+    # swap lo/hi if the derivative points to a narrower bracket width
+    in_bracket = lambda x, y: ((x < y) & (y < 0) | (x > y) & (y > 0))
+    swap_lo_next = in_bracket(lo.deriv_0, lo_next.deriv_0)
+    lo = jax.tree_util.tree_map(
+        lambda x, y: jp.where(swap_lo_next, y, x), lo, lo_next
+    )
+    swap_lo_mid = in_bracket(lo.deriv_0, mid.deriv_0)
+    lo = jax.tree_util.tree_map(
+        lambda x, y: jp.where(swap_lo_mid, y, x), lo, mid
+    )
+    swap_lo_hi_next = in_bracket(lo.deriv_0, hi_next.deriv_0)
+    lo = jax.tree_util.tree_map(
+        lambda x, y: jp.where(swap_lo_hi_next, y, x), lo, hi_next
+    )
+    swap_hi_next = in_bracket(hi.deriv_0, hi_next.deriv_0)
+    hi = jax.tree_util.tree_map(
+        lambda x, y: jp.where(swap_hi_next, y, x), hi, hi_next
+    )
+    swap_hi_mid = in_bracket(hi.deriv_0, mid.deriv_0)
+    hi = jax.tree_util.tree_map(
+        lambda x, y: jp.where(swap_hi_mid, y, x), hi, mid
+    )
+    swap_hi_lo_next = in_bracket(hi.deriv_0, lo_next.deriv_0)
+    hi = jax.tree_util.tree_map(
+        lambda x, y: jp.where(swap_hi_lo_next, y, x), hi, lo_next
+    )
+    swap = swap_lo_next | swap_lo_mid | swap_lo_hi_next
+    swap = swap | swap_hi_next | swap_hi_mid | swap_hi_lo_next
     ctx = ctx.replace(lo=lo, hi=hi, swap=swap, ls_iter=ctx.ls_iter + 1)
 
     return ctx
@@ -331,14 +502,17 @@ def solve(m: Model, d: Data) -> Data:
   def body(ctx: _Context) -> _Context:
     ctx = _linesearch(m, d, ctx)
     prev_grad, prev_Mgrad = ctx.grad, ctx.Mgrad  # pylint: disable=invalid-name
-    ctx = _update_constraint(d, ctx)
+    ctx = _update_constraint(m, d, ctx)
     ctx = _update_gradient(m, d, ctx)
 
-    # polak-ribiere:
-    beta = jp.dot(ctx.grad, ctx.Mgrad - prev_Mgrad)
-    beta = beta / jp.maximum(mujoco.mjMINVAL, jp.dot(prev_grad, prev_Mgrad))
-    beta = jp.maximum(0, beta)
-    search = -ctx.Mgrad + beta * ctx.search
+    if m.opt.solver == SolverType.NEWTON:
+      search = -ctx.Mgrad
+    else:
+      # polak-ribiere:
+      beta = jp.dot(ctx.grad, ctx.Mgrad - prev_Mgrad)
+      beta = beta / jp.maximum(mujoco.mjMINVAL, jp.dot(prev_grad, prev_Mgrad))
+      beta = jp.maximum(0, beta)
+      search = -ctx.Mgrad + beta * ctx.search
     ctx = ctx.replace(search=search, solver_niter=ctx.solver_niter + 1)
 
     return ctx